// frontend/src/api/generated/newChatbotAPI.schemas.ts
export const GetAdminCoupons200AllOfTwoDataItemDiscountType = {  PERCENTAGE: 'PERCENTAGE',
  FIXED: 'FIXED',
} as const
export type GetAdminCoupons200AllOfTwoDataItem = {
  id: string;
  code: string;
  discountType: typeof GetAdminCoupons200AllOfTwoDataItemDiscountType[keyof typeof GetAdminCoupons200AllOfTwoDataItemDiscountType] ;
  value: number;
  expiresAt?: string;
  usageLimit?: number;
  usageCount: number;
  active: boolean;
  stripeCouponId?: string;
  createdAt: string;
  updatedAt: string;
};